+++ BreadPlayer.Core/Engines/BASSEngine/BASSPlayerEngine.cs
@@ -57,15 +57,16 @@ namespace BreadPlayer.Core.Engines.BASSEngine
                 {
                     Bass.UpdatePeriod = 230;
                     Bass.Start();
-                    //we set it to a high value so that there are no cuts and breaks in the audio when the app is in background.
-                    //This produces latency issue. When pausing a song, it will take 230ms. But I am sure, we can find a way around this later. 
+
                     if (isMobile)
                     {
+                        //we set it to a high value so that there are no cuts and breaks in the audio when the app is in background.
+                        //This produces latency issue. When pausing a song, it will take 230ms. But I am sure, we can find a way around this later. 
                         NativeMethods.BASS_SetConfig(NativeMethods.BassConfigDevBuffer, 230);
-                    }
-
+                    }                   
+                    Bass.Configure(Configuration.IncludeDefaultDevice, true);
                     Bass.Init();
-                    Effect = new Effects();
+                    //Effect = new Effects();
                 }
                 catch(Exception)
                 {
@@ -115,13 +116,20 @@ namespace BreadPlayer.Core.Engines.BASSEngine
                         Length = 0;
                         Length = Bass.ChannelBytes2Seconds(_handle, Bass.ChannelGetLength(_handle));
                         Bass.FloatingPointDSP = true;
-                        Effect.UpdateHandle(_handle);
                         Bass.ChannelSetSync(_handle, SyncFlags.End | SyncFlags.Mixtime, 0, _sync);
                         Bass.ChannelSetSync(_handle, SyncFlags.Position, Bass.ChannelSeconds2Bytes(_handle, Length - 5), _posSync);
                         Bass.ChannelSetSync(_handle, SyncFlags.Position, Bass.ChannelSeconds2Bytes(_handle, Length - 15), _posSync);
                        
                         CurrentlyPlayingFile = mediaFile;
                     });
+                    if (Equalizer == null)
+                    {
+                        Equalizer = new BassEqualizer(_handle);
+                    }
+                    else
+                    {
+                        (Equalizer as BassEqualizer).ReInit(_handle);
+                    }
                     MediaStateChanged?.Invoke(this, new MediaStateChangedEventArgs(PlayerState.Stopped));
 
                     return true;
@@ -157,8 +165,8 @@ namespace BreadPlayer.Core.Engines.BASSEngine
             MediaStateChanged?.Invoke(this, new MediaStateChangedEventArgs(PlayerState.Paused));
             await Task.Run(async () =>
             {
-                Bass.ChannelSlideAttribute(_handle, ChannelAttribute.Volume, 0, 700);
-                await Task.Delay(700);
+                Bass.ChannelSlideAttribute(_handle, ChannelAttribute.Volume, 0, 500);
+                await Task.Delay(500);
                 Bass.ChannelPause(_handle);
                 //var vol = (float)Volume / 100f;
                 //Bass.ChannelSetAttribute(handle, ChannelAttribute.Volume, vol);
@@ -176,7 +184,7 @@ namespace BreadPlayer.Core.Engines.BASSEngine
                 Bass.ChannelSetAttribute(_handle, ChannelAttribute.Volume, 0f);
                 Bass.ChannelPlay(_handle);
                 var vol = (float)Volume / 100f;
-                Bass.ChannelSlideAttribute(_handle, ChannelAttribute.Volume, vol, 3000);
+                Bass.ChannelSlideAttribute(_handle, ChannelAttribute.Volume, vol, 1000);
                 PlayerState = PlayerState.Playing;
             });
             MediaStateChanged?.Invoke(this, new MediaStateChangedEventArgs(PlayerState.Playing));
@@ -287,9 +295,22 @@ namespace BreadPlayer.Core.Engines.BASSEngine
             set => Set(ref _ignoreErrors, value);
         }
 
-        public Equalizer Equalizer { get => throw new NotImplementedException(); set => throw new NotImplementedException(); }
-        public bool IsLoopingEnabled { get => throw new NotImplementedException(); set => throw new NotImplementedException(); }
-
+        private Equalizer _fmodEqualizer;
+        public Equalizer Equalizer
+        {
+            get => _fmodEqualizer;
+            set => Set(ref _fmodEqualizer, value);
+        }
+        private bool _isLoopingEnabled;
+        public bool IsLoopingEnabled
+        {
+            get => _isLoopingEnabled;
+            set
+            {
+                Set(ref _isLoopingEnabled, value);
+                //SetLoop();
+            }
+        }
         #endregion
         private void PositonReachedSync(int handle, int channel, int data, IntPtr user)
         {