+++ BreadPlayer.Core/Engines/Interfaces/IEqualizer.cs
@@ -54,10 +54,17 @@ namespace BreadPlayer.Core.Engines.Interfaces
                     Init();
                 }
                 else
-                {                    
+                {
+                    SaveEqualizerSettings();
                     DeInit();
                 }
             }
+        }
+        private float _preamp = 1f;
+        public float Preamp
+        {
+            get => _preamp;
+            set => _preamp = value;
         }
         public bool IsPreampAvailable { get; set; }
         public string Name { get; set; }