+++ BreadPlayer.Core/Models/Playlist.cs
@@ -47,6 +47,18 @@ namespace BreadPlayer.Core.Models
         public string Salt { get => _salt;
             set => Set(ref _salt, value);
         }
+        private bool _isExternal;
+        public bool IsExternal
+        {
+            get => _isExternal;
+            set => Set(ref _isExternal, value);
+        }
+        private string _path;
+        public string Path
+        {
+            get => _path;
+            set => Set(ref _path, value);
+        }
         public List<long> SongsIds { get; set; } = new List<long>();
 
         public string GetTextSearchKey()