+++ BreadPlayer.Database/IDatabaseService.cs
@@ -17,7 +17,8 @@ namespace BreadPlayer.Database
         T GetRecordById<T>(long id);
         Task RemoveRecords(IEnumerable<IDbRecord> records);
         Task RemoveRecord(IDbRecord record);
-        void UpdateRecords(IEnumerable<IDbRecord> records);
+
+        void UpdateRecords<T>(IEnumerable<IDbRecord> records);
         Task<bool> UpdateRecordAsync<T>(T record, long id);
         Task<IEnumerable<T>> QueryRecords<T>(string term);
         int GetRecordsCount();