+++ BreadPlayer.Views.UWP/Shell.xaml.cs
@@ -31,6 +31,8 @@ using BreadPlayer.Helpers;
 using BreadPlayer.Messengers;
 using BreadPlayer.ViewModels;
 using SplitViewMenu;
+using BreadPlayer.Core.Common;
+using Windows.UI.ViewManagement;
 
 namespace BreadPlayer
 {
@@ -54,28 +56,51 @@ namespace BreadPlayer
                 Tooltip = "Enable Multiselection",
                 ShortcutCommand = (Application.Current.Resources["LibVM"] as LibraryViewModel).ChangeSelectionModeCommand
             });
-            NowPlayingItem.Command = _shellVm.NavigateToNowPlayingViewCommand;
-          
+            NowPlayingItem.Command = new DelegateCommand(() => 
+            {
+                _shellVm.IsPlaybarHidden = true;
+                ApplicationView.GetForCurrentView().TryEnterFullScreenMode();
+            });          
         }
-           
+        private void VisualStateGroup_CurrentStateChanged(object sender, VisualStateChangedEventArgs e)
+        {
+            if (NowPlayingFrame.CurrentSourcePageType != typeof(NowPlayingView))
+                NowPlayingFrame.Navigate(typeof(NowPlayingView));
+        }
         protected async override void OnNavigatedTo(NavigationEventArgs e)
         {
             Window.Current.CoreWindow.KeyDown += (sender, args) =>
             GlobalPageKeyDown?.Invoke(sender, args);
             if (RoamingSettingsHelper.GetSetting<bool>("IsFirstTime", true))
             {
-                string releaseNotes = "FIXES:\r\n\r\n" + 
-                    "We fixed the startup crash.\n" +
-                    "Fixed crash when changing theme.\n" +
-                    "Fixed crash when play songs from an album.\n" +
-                    "Fixed other various bugs and crashes.\r\n\r\n" + 
+                string releaseNotes = "FIXES:\r\n\r\n" +
+                    "Fixed issue with output device not changing when headphones are connected.\n" +
+                    "Fixed NowPlayingList loses its ItemTemplate when window size is changed.\n" +
+                    "Fixed crash after library load.\n" +
+                    "Fixed playlist import.\n" +
+                    "Fixed audio stutter when minimizing to background (only mobile).\n"+
+                    "Fixed crash when deleting duplicates.\r\n\r\n" + 
                     "NEW THINGS:\r\n\r\n" +
-                    "Added 'Now Playing' Screen with artist and album info fetching.\r\n\r\n" +
+                    "Added new Keyboard Shortcuts ('Keybindings' section in Settings).\n" + 
+                    "Added full translation support.\n" +
+                    "Added ability to export playlist (only to .m3u & .pls for now).\n" +
+                    "Added silent upcoming song notifications.\n" + 
+                    "Added translations for Czech and Sinhala languages.\n" +
+                    "Added equalizer presets (experimental).\n" +
+                    "Added real-time music library updates when filesystem changes (only works when app is running).\n" +
+                    "Added 'Contribute' section in settings to help in contributing.\n" +
+                    "Added auto playback stop in BreadPlayer when a song is played in Groove Player.\n" +
+                    "Added navigate to now playing screen when song is played on mobile.\n" +
+                    "Added auto removal of duplicates when importing songs.\n"+
+                    "Added navigate to now playing screen when song's tags are clicked in mini player.\r\n\r\n" +
                     "IMPROVEMENTS:\r\n\r\n" +
-                    "Greatly improved performance and navigation.\n" +
-                    "Improved Equalizer UI.\nImproved Blur\n" +
-                    "Improved UI (less glitches etc.)\r\n";
-                await SharedLogic.NotificationManager.ShowMessageBoxAsync(releaseNotes, "What's new in v2.3.0");
+                    "Improved hamburger menu list item response on tap/click.\n" +
+                    "Improved all animations and transitions.\n" +
+                    "Improved navigation performance.\n"+
+                    "Improved library import.\n" + 
+                    "Improved overall UI and increased readability.\n" +
+                    "Improved startup performance by 50% (only PC).\r\n";
+                await SharedLogic.NotificationManager.ShowMessageBoxAsync(releaseNotes, "What's new in v2.3.0 (after 120+ commits)");
                 RoamingSettingsHelper.SaveSetting("IsFirstTime", false);
             }
             if (e.Parameter is StorageFile)
@@ -139,6 +164,8 @@ namespace BreadPlayer
                 _shellVm.DontUpdatePosition = true;
                 _isProgBarPressed = true;
             }
-        }           
+        }
+
+      
     }
 }